Vector backends now return Documents
